Include item date and item status date in item json
The round-trip is cleaner with the dates, and in the case of recently
returned materials, the public interface should be able to report when
material was returned. I am, however, reporting the dates in the Json as
seconds-since-the-epoch, rather than milliseconds-since-the-epoch. The
latter is standard in Java, but not in Ruby, and milliseconds are not
part of the timestamp information from Oracle so reporting milliseconds
as zeroes every time would be both pointless and inconvenient for the
front end.