More work on holdings, add BoundWith support
The bound with implementation is intended to pull information currently
in the Solr index as well as the live bound-with data pulled through the
current availability service. It doesn't include the holdings id for the
current record because it will eventually be include as subsidiary data
under the holding's json.

The holdings work is a lot more complete, but currently needs bound-with
integration, bound-with blocking on empty items, and various non-json
fields, particularly boolean filter flags.