Modify JUnit tests to use static test database
The new PopulateTestVoyagerDB creates (or adds to) a SQLite database
file that can be included in the repo and will support all of the
existing tests. I'm going to hold off on committing the database file
itself until it's a little more stable, as it's a binary file.

In addition to test stability (no failures because a book is checked
out) and portability (no need to access Cornell's Voyager Oracle DB),
this is intended to make it easier to build tests around complex and
changing availability states.