Filter out suppressed holdings; Add to tests
We now have an items/holdings test for a missing item, and a
holdings-only test for a 309 item serial with one item apparently at
bindery since 2015. In the latter example, I'm concerned about the
source of the note in the catalog that reads, "Sent to bindery for 1
month". Wherever that acomes from, I probably need to load it.