Identify items not in the location as the rest of the holding
When some but not all of the items associated with a holding record are
in a temporary location, then instead of overriding the holdings display
location we need to handle the mixed location for that holding record.
This new example introduces ${mfhdid}->items->tempLoc as an array of
items not in the location of the main holdings. This is not mutually
exclusive of the same item being referenced in
${mfhdid}->items->unavail.