Use status ranking to select top status per item
The status "ranking" is not in the database with the status types
themselves, but in the circulation documentation. This makes for an
awkward and fragile data join, but it's likely that this map won't
change during the remaining lifetime of the system. The most likely
change is probably that someone decides to modify the display name of
one of the statuses, in which case it will be ranked at the bottom of
the list until the code is updated to match.