Add mapping from location to location facet values
LocationFacetMapping.txt is a tab-delimited file saved from an Excel spreadsheet. The spreadsheet is intended to give library staff a comfortable way to both review and maintain the mapping. The mapping works according to first-matching-rule. Suppressed rules will not appear in the facet and will not allow later rules to be evaluated for a location.

This implementation creates a new point of failure, requiring that the spreadsheet be updated to match any changed shelving locations. If this doesn't happen, the current version will simply complain to STDOUT, but once enough testing has occurred a stronger reporting mechanism will be appropriate for production.