Better configure for command-line Maven execution
Note that the inclusion of the dependency on ojdbc as a runtime scope dependency instead of a sytem scope dependency requires that the library be installed before using. The command for this is:

mvn install:install-file -DgroupId=com.oracle -DartifactId=ojdbc7 -Dversion=12.1.0.2 -Dpackaging=jar -Dfile=lib/ojdbc7-12.1.0.2.jar -DgeneratePom=true

This should probably be configured into an appropriate lifecycle phase in the pom.xml, but I'll deal with that later if it's an issue.