Finish inventory monitoring and change tracking
The new code will not load a document to Solr that doesn't represent the expected change, will continue to expect the change until it appears, and will load it to Solr only once when the change does appear. If a change is loaded to Solr incidentally because of a different update before it is expected, then it will load a second time once the change is expected.