Forgive 1 second timestamp discrepancies in record dates
Because record dates used for particular Solr document versions are tracked using the dates internal to the MARC records rather than those in the Voyager DB datetime columns, discrepancies can happen. As a result, the code may refuse to accept a Solr document update as being the expected update because the Solr document appears to be built with a version of the MARC a second older than the expected MARC. This is an illusion, and the long-term solution should be to depend on only a single timestamp date records. The short-term solution implemented here, is to accept the one-second-too-old data after three minutes have elapsed, during which time there should be time to have incorporated the second update if the two timestamps do represent separate record changes.