Clean up launch window logic
Rather than using a hard-coded start time for the monitoring apps to use when no start time is found in the db, they now default to starting an update sweep starting 10 hours before the current time. It's not intended that this default will play a significant role.

Also, FindBugs is happier with the method RecordsToSolr.getCurrentToDate(), which is cleaner now that it no longer passes through the default start time.