Move Call Number analysis to custom filter; create callnum browse Solr config
DISCOVERYACCESS-19
This is testing well, but still needs to be run past UX to confirm the list of display fields and facet fields. If needed, some of these fields may need to be broken down at a lower level granularity, as the fields are currently granular to the bib, while fields like location and availability may apply to select call numbers on the bib, while some call numbers may be bibliographic and are not connected with particular location/availability information.

The current query syntax requires the main query be tagged for exclusion from faceting, e.g. "/solr/corename/browse?q={!tag=mq}["DS561" TO *]"

More detailed usage info in Jira.