Update call number filter to recognize configurable prefixes
DISCOVERYACCESS-19

This will make it easy to supply the call numbers with prefixes to the index, as well as accept searches with prefixes from patrons, all while sorting according to the call number sans prefex. The side effect is that the prefixes are not, themselves, browseable, but they will still be searchable in the general call number search.