Populate online facet in call number browse based on specific holdings
Rather than copy over the field from the blacklight index records, the call number browse index should have this value generated by the specific holdings. For most records, the results will be the same, but when a bib represents both print and online holdings, the call numbers on the records may each represent either one or the other.

DISCOVERYACCESS-19