Suppress non-call numbers from call# browse; Add shelfloc bool
The specific call number values "No Call Number", "In Process", and "On Order" can be safely suppressed when no better call number is available. The shelfloc boolean is only populated when true, and should be populated when the physical holdings are represented, and the call number in question comes from the holding record(s). It may also be a good idea to suppress this boolean based on location, where many locations don't allow patrons in the stacks, and in the case of the annex don't allow patrons in the stacks and also don't store the books by call number.

DISCOVERYACCESS-19