If serv,remo holdings but no access link, omit from call num browse
DISCOVERYACCESS-19

The serv,remo holdings should indicate that the record has online holdings, so the lack of an access link suggests that either the cataloging is incorrect, or the indexing job as misclassified an access link as some other kind of link. Because there's an error somewhere, we'll also log the problem to STDOUT.

Also, change the default output format from the Solr call number index to json.