Detect changes to item statuses in reserves module
Most of our libraries don't use the Voyager reserves module, but just edit item records to add (or remove) a temporary location and loan type. The reserves module, when it is used, makes these changes without updating the modification date on the item records. So to detect these changes, we need to start monitoring timestamps in the reserves module database tables instead.

This is most of the implementation, but the implementation needs to prove up, needs tests, and is currently using a hard-coded time stamp in Items.java.

DISCOVERYACCESS-4494