Restructure RefreshAvailability to avoid double processing
The delay in indexing means that immediately after submitting a record to Solr, it may still appear to be one of the least recently touched records for a time. In testing, this meant often re-indexing the same record 2, 3, 4, or occassionally 5 times before the indexing caught up and reflected the change. A little bit of cyclical paging in the "oldest record" query should ensure that we don't hit a record again during its update delay window.