Drop fields from call number browse index not selected by reps
DISCOVERYACCESS-19
At the reps' meeting today, it was decided that the facets we want are location, online, language, pub date, and acquisition date. The only metadata field to be added to the main display is format. Some of the other facets are large and expensive to include in the index, so I'm dropping them now. If any are added to the spec later, they can be re-added.