Create task to confirm and update all recent serial issues in Solr
Rather than adding a Solr field to track records with currently displaying serial issues, I've created a database table to track them. This means that it will be important to take steps that a new version of a Solr index always has its own version of this table, as there's a built-in assumption that the table reflects accurately what a given Solr index contains w/r/t serial issues. The benefit is that this full sweep can be run with greater frequency, comparing the contents of the table to the data in Voyager, and only involving Solr where discrepancies are found.

DISCOVERYACCESS-4253