Guard against item deleted immediately after status change
This is probably an extremely rare condition, but it should be protected against. It seems that the item (not identified in the log) had a status change and was then deleted quickly enough that while the changed status was detected by detectChangedItemStatuses(), the item was no longer present to retrieve details about.

For the time being, I'll log and otherwise ignore cases like this, as the item delete should be otherwise detectable.

java.lang.NullPointerException
    at edu.cornell.library.integration.voyager.Items.detectChangedItemStatuses (Items.java:114)
    at edu.cornell.library.integration.availability.MonitorAvailability.main (MonitorAvailability.java:56)
    at sun.reflect.NativeMethodAccessorImpl.invoke0 (Native Method)
    at sun.reflect.NativeMethodAccessorImpl.invoke (NativeMethodAccessorImpl.java:62)
    at sun.reflect.DelegatingMethodAccessorImpl.invoke (DelegatingMethodAccessorImpl.java:43)
    at java.lang.reflect.Method.invoke (Method.java:498)
    at org.codehaus.mojo.exec.ExecJavaMojo$1.run (ExecJavaMojo.java:282)
    at java.lang.Thread.run (Thread.java:745)