Improve display of time-since duration to support wider time ranges
Usually, we hope to report that a change in the catalog made it to the OPAC within a handful of seconds, but some changes, such as re-indexing things that haven't been indexed recently get into the range of days. With the new behavior, elapsed time can be represented in days, hours, minutes, or seconds, with only the most significant non-zero unit and the one just under it reported. That is, if a duration reports a time in days, then hours may also report, but minutes and seconds do not.