Add item record change detection to availability monitor
DISCOVERYACCESS-4886

The item record changes are still being detected in the record change monitor as well, which means that during daytime processing, item changes will be detected twice. I'm going to leave that as is for now, however, because the daytime record change detection has the ability to queue updates for two separate bibs when an item is reassigned from one holding record to another. This can't be easily moved to the availability monitor, as the availability monitor operates without knowledge of the pre-existing item inventory.