Trigger reindex of included bibs when re-indexing bound-with master items
DISCOVERYACCESS-4540

Every time we reindex a bib with bound-with master items attached, we'll reindex any bibs that may be included in those items. In the future, this logic could be tightened to trigger reindex only when the items in question have actually changed status, and only those bibs included in the particular items. (Some high level bibs contain a good number of such master items.)