Update Solr Json to reflect Annex nocirc items unavailable
When the item status on these items is a theoretically available status, we will replace it for display with the generic "Unvailable" status. This means that other item statuses will not be lost in display, but the material that patrons have no access to won't appear available regardless of status.

DISCOVERYACCESS-4917
DISCOVERYACCESS-4881