Increase priority of bound-with knock-on reindexing
In the original implementation, we were doing knock-on re-indexing very generously, which meant a lower priority was appropriate, but was problematic when circular relationships caused infinite re-indexing loops. Now that priority 8 is being used for age-of-record reindexes as well, using this priority for bound-with knock-ons as well may mean these don't always get through the queue efficiently.

DISCOVERYACCESS-4540