Make item change detection work with "Unavailable" nocirc annex items
We recently, in service of DISCOVERYACCESS adjusted the Solr availability representation for theoretically "Available" items that are not actually available for public use. As part of that, I removed the status date from the Solr representation entirely to keep it from displaying meaninglessly to patrons. This tripped up the process that detects status changes, as a timestamp associated with the change is a necessary component of queueing and logging the detected change. So, we'll just use the current timestamp as a stand-in for this Very Infrequent case. (Materials not available to the public don't circulate much.)