Add order detection for pending orders
This implements the second bullet point in DISCOVERYACCESS-4906.

Pending orders will be displayed as "In pre-order processing as of [date]", using the line item copy status date, or simply "In pre-order processing" when the line item copy status is undated.

Also, the display of order dates has been switched to the m/d/yy format that is the default for unavailable item status and due date displays, and the unit tests for OpenOrder are now exclusively using the canned test sqlite database rather than the live catalog database.