Improve call number browse for available-to-purchase titles
Enter the available-to-purchase titles under their bibliographic record call numbers instead of the dummy holdings call number, "Available for the Library to Purchase". Also, put the "Available for the Library to Purchase"  in the bibliographic summary in place of the book's location, so it shows in the availability column in the browse.

DISCOVERYACCESS-4872