Pull multiple open orders for a bib
When multiple copies of a book are on order, we want to pull the order information for all of them. These can be a single line item for two copies with two line item copy records associated with different holdings, one line item with two line item copies incorrectly associated with the same holding record, or two different orders entirely, which may be mistakenly associate with a single holding record.

The high rate of misassociation between line item copies and their respective holding records derives from the acquisitions module downplaying the difference between holdings in the same location, only asking the staff to identify the intended location for the ordered book. While it is possible to get the assocations correct, it's very easy to get them incorrect, and the associations can no longer be corrected once the order is approved.

DISCOVERYACCESS-4906