Adjust json serialization
Had an issue testing the implementation of DISCOVERYACCESS-4873 because of a change in the json serialization behavior in the 2.9.9 jackson library. (DISCOVERYACCESS-4970). Specifically, the serialization now interprets the NON_EMPTY directive to drop key value pairs from a serialized hash when the value is "empty", even if the key is not, leading to dropping the hash entirely when all the values are "empty". (Why we need a hash when none of the values are populated is a story, but that's the situation here.)  I've switched to NON_NULL serialization, and am now more agressively using nulls when I don't want things to serialize.