Don't manually load Mysql driver class
This is no longer necessary, and apparently with the more recent mysql client it's no longer the correct class name, either.

Loading class `com.mysql.jdbc.Driver'. This is deprecated. The new driver class is `com.mysql.cj.jdbc.Driver'. The driver is automatically registered via the SPI and manual loading of the driver class is generally unnecessary.