Don't use Timestamp#toLocalDateTime
I was using this method to get a DateTime object to pass to the DateTimeFormatter. This method does two things, however. It creates a DateTime object in the local time zone and also adjusts the time in the Timestamp object from UTC to the local timezone. This is generally a bad idea because Timestamps don't internally encode their timezones and it becomes difficult to track if their internal timezone isn't consistent.