Dedupe recent issues for display
There's a particular suppressed bib that has 203,522 received issues, all flagged for display in OPAC, and all with the exact same name. This isn't ever going to display, since the bib is suppressed, but the json representation is large enough to not fit in a "text" column of a mysql db. Rather than widen the field to support this non-display display data, I've added a dedupe.
DISCOVERYACCESS-5823