Tighten up conditions for batch processing deprioritization
After watching this run for a while, the solution seems quite effective in identifying batch processing and reducing its processing priority. The threshold of 10 updates in 5 minutes does sometimes capture manual work and integrations work. Integrations work (like requests, online renewals, gobi processing) can be vulnerable because it runs through shared user accounts. It doesn't seem to be doing any harm since the rate of non-batch processing is low enough, that applying the stronger deprioritization to actual batch processing is enough to avoid noticeable lag. But I'm going to tighten the minimum threshold for deprioritization to 25 updates in 3 minutes instead of 10 updates in 5.
DISCOVERYACCESS-7307