Solr rejecting call number browse record results in infinite loop
We're already processing one bib at a time through updateBib(s)InSolr(), but by removing the loop that implies we could be doing more than one we can have a success boolean as a return value. If the update was unsuccessful we can avoid leave the record on the queue but not remove the bibLock on it, so it won't immediately be re-selected for processing.
DISCOVERYACCESS-7437