Tie bib-level POD status tracking to the instance rather than the MARC
Because we're tracking the POD status of non-MARC Folio instances (never delivered to POD as we don't have support for non-MARC bibs at this time), their absence from the marc Folio cache makes them look like they've been deleted from Folio. It's not a major problem, as it probably adds only about a minute to evaluate all 16k+ non-MARC instances and to decide no POD update is needed... but we may see dramatic increases in non-MARC record counts in the future.
DISCOVERYACCESS-6587
RES-8