Use item sequence number from itemSequence table
The REPLACE INTO command I'm using to update the item records in the itemFolio table resets the itemFolio.sequence field to null. I could address this by retrieving the value first and setting it as part of the REPLACE statement, but this seems cleaner. The other change would need to take place in both the live change monitor and the overnight LDP change detection process.
DISCOVERYACCESS-7514