Increase logging detail on incremental POD export case where bib marc updated
I'm seeing false positives in the log, where records are exported to POD in the incrementals because their bib marc is "newer", despite nothing having changed. A moderate rate of false positives is acceptable, but a similar rate of false negatives would be more damaging. So I'm increasing the logging detail to better understand what's happening.

RES-1
RES-8
DISCOVERYACCESS-6587