Add modification dates to POD Incremental export log when not pushing an update
It looks like some records are being resent to POD because the current record date contains milliseconds and our tabulation of the date on the version last sent only contains seconds. As a result, the milliseconds make the record look "newer". I'm curious now about all the records that are not sent for this reason, as the odds of milliseconds not making the record look newer should be 1 in 1,000.

RES-1
RES-8
DISCOVERYACCESS-6587