Use folio cache mysql moddate column for bib marc
When writing the moddate from the json to the db column at the time of caching the record, MySQL rounds the timestamp to the nearest second, as its timestamp column doesn't support microseconds. By comparing the unrounded moddate directly from the json with the rounded moddate in the POD inventory tables, the bib marc appears to be newer or older than the bib marc last sent to POD by microseconds, depending on the direction of the rounding. By using the value as rounded by MySQL, we're using two values both rounded by the same system in the same way and should get a much more consistent comparison.

RES-8
DISCOVERYACCESS-6587