Make POD upload text closer to curl's
The attempt to upload incremental POD updates using the java function returned a 400 error from the POD server. Looking at the structure of the multipart/form-data produced when uploading with curl, a few discrepancies appear. A missing extra pair of brackets in upload[files][], a missing repetition of the filename on that same line, and the syntax for identifying the content type of the uploaded file. Fingers crossed!!

RES-8
DISCOVERYACCESS-6587