Further tweaks to POD push
The changes are still yielding a 400 error from the POD server. I'm going to try not putting the multipart boundry in quotes in the HTTP Content-Type property. The spec includes the quotes, but curl doesn't appear to use them, so they're at least unnecessary. I'm also borrowing a boundary value from curl. Using a random UUID as the boundary was a trick I found online, but just in case the syntax is problematic, I'll try the hardcoded value.

RES-8
DISCOVERYACCESS-6587