Move POD incremental date cursor forward
This won't effect the result of the incremental generation, but I want to see how much faster the generation runs when considering a narrower date range. It's prohibitively slow when the cursor is set to pre-Folio go-live, but ~5 wks record updates is probably making very little difference. If I'm wrong, I might want to build a proper sliding cursor.