Fix bib loop continuations
Specifically, the NoEx continue was applied to a lesser loop in the code and the NoEx records were not skipped. The majority appear to have been skipped anyway due to being e-resources and having no items. So I've cleaned up the looping and moved the NoEx check to after the item check to get better numbers on the records actually excluded for NoEx.

I've also updated the query in getBibsToExport() to include the non-MARC instances so we can see numbers on those as they're eliminated.

DISCOVERYACCESS-7984